[conf] Rename dataBinding to br-inline.
