+++ booster-transform-br-inline/README.md
@@ -1,4 +1,4 @@
-# booster-transform-databinding
+# booster-transform-br-inline
 
 This module is used for BR index inline if databinding is enabled.
