+++ booster-transform-br-inline/src/test/resources/BR.class
