+++ settings.gradle
@@ -30,7 +30,7 @@ include ':booster-task-spi'
 include ':booster-transform-activity-thread'
 include ':booster-transform-all'
 include ':booster-transform-asm'
-include ':booster-transform-databinding'
+include ':booster-transform-br-inline'
 include ':booster-transform-finalizer-watchdog-daemon'
 include ':booster-transform-javassist'
 include ':booster-transform-lint'
